Application: porting to Linux using GLFW3
finally decided on using this library, since I
can't really avoid an external dependency here

missing gamepad support since I don't have
access to GLFW3.3 at the moment

+++ Source/Runtime/Core/Public/Maths/MathHelpers-inl.h
@@ -37,7 +37,7 @@ inline CONSTEXPR bool Lerp(bool v0, bool v1, float f) NOEXCEPT {
 template <typename T, class>
 CONSTEXPR float LinearStep(T value, Meta::TDontDeduce<T> vmin, Meta::TDontDeduce<T> vmax) NOEXCEPT {
     Assert(vmin < vmax);
-    return Saturate((value - vmin) / static_cast<float>(vmax - vmin));
+    return Saturate(static_cast<float>((value - vmin) / static_cast<float>(vmax - vmin)));
 }
 //----------------------------------------------------------------------------
 template <typename T, class>
